+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-# Changes
+# CHANGELOG
 
+---
+- REL : Version `0.0.6`
+- ENH : Generate `.d.ts` for js bundled files
+- ENH : Add `ds-bundle` and creating plugin for bundler `.d.ts` files
+
+---
 - REL : Version `0.0.5`
 - OTH : Create chunks with webpack
 - OTH : Change some exports default